Move `KeySize` to its own file; export and document it
