+++ crates/table/src/btree_index.rs
@@ -25,15 +25,16 @@ use super::indexes::RowPointer;
 use super::table::RowRef;
 use crate::{read_column::ReadColumn, static_assert_size, MemoryUsage};
 use core::ops::RangeBounds;
-use spacetimedb_lib::ProductValue;
 use spacetimedb_primitives::{ColList, IndexId};
 use spacetimedb_sats::{
-    algebraic_value::Packed, i256, product_value::InvalidFieldError, u256, AlgebraicType, AlgebraicValue, ArrayValue,
-    ProductType, SumValue,
+    algebraic_value::Packed, i256, product_value::InvalidFieldError, u256, AlgebraicType, AlgebraicValue, ProductType,
 };
 
+mod key_size;
 mod multimap;
 mod uniquemap;
+
+pub use key_size::KeySize;
 
 type Index<K> = multimap::MultiMap<K, RowPointer>;
 type IndexIter<'a, K> = multimap::MultiMapRangeIter<'a, K, RowPointer>;
@@ -657,124 +658,6 @@ impl TypedIndex {
     }
 }
 
-trait KeySize {
-    fn key_size_in_bytes(&self) -> usize;
-}
-
-macro_rules! impl_key_size_primitive {
-    ($prim:ty) => {
-        impl KeySize for $prim {
-            fn key_size_in_bytes(&self) -> usize { std::mem::size_of::<Self>() }
-        }
-    };
-    ($($prim:ty,)*) => {
-        $(impl_key_size_primitive!($prim);)*
-    };
-}
-
-impl_key_size_primitive!(
-    bool,
-    u8,
-    i8,
-    u16,
-    i16,
-    u32,
-    i32,
-    u64,
-    i64,
-    u128,
-    i128,
-    spacetimedb_sats::algebraic_value::Packed<u128>,
-    spacetimedb_sats::algebraic_value::Packed<i128>,
-    u256,
-    i256,
-    spacetimedb_sats::F32,
-    spacetimedb_sats::F64,
-);
-
-impl KeySize for Box<str> {
-    fn key_size_in_bytes(&self) -> usize {
-        self.len() + std::mem::size_of::<super::var_len::VarLenRef>()
-    }
-}
-
-impl KeySize for AlgebraicValue {
-    fn key_size_in_bytes(&self) -> usize {
-        match self {
-            AlgebraicValue::Bool(x) => x.key_size_in_bytes(),
-            AlgebraicValue::U8(x) => x.key_size_in_bytes(),
-            AlgebraicValue::I8(x) => x.key_size_in_bytes(),
-            AlgebraicValue::U16(x) => x.key_size_in_bytes(),
-            AlgebraicValue::I16(x) => x.key_size_in_bytes(),
-            AlgebraicValue::U32(x) => x.key_size_in_bytes(),
-            AlgebraicValue::I32(x) => x.key_size_in_bytes(),
-            AlgebraicValue::U64(x) => x.key_size_in_bytes(),
-            AlgebraicValue::I64(x) => x.key_size_in_bytes(),
-            AlgebraicValue::U128(x) => x.key_size_in_bytes(),
-            AlgebraicValue::I128(x) => x.key_size_in_bytes(),
-            AlgebraicValue::U256(x) => x.key_size_in_bytes(),
-            AlgebraicValue::I256(x) => x.key_size_in_bytes(),
-            AlgebraicValue::F32(x) => x.key_size_in_bytes(),
-            AlgebraicValue::F64(x) => x.key_size_in_bytes(),
-            AlgebraicValue::String(x) => x.key_size_in_bytes(),
-            AlgebraicValue::Sum(x) => x.key_size_in_bytes(),
-            AlgebraicValue::Product(x) => x.key_size_in_bytes(),
-            AlgebraicValue::Array(x) => x.key_size_in_bytes(),
-
-            AlgebraicValue::Min | AlgebraicValue::Max => unreachable!(),
-        }
-    }
-}
-
-impl KeySize for SumValue {
-    fn key_size_in_bytes(&self) -> usize {
-        1 + self.value.key_size_in_bytes()
-    }
-}
-
-impl KeySize for ProductValue {
-    fn key_size_in_bytes(&self) -> usize {
-        self.elements.key_size_in_bytes()
-    }
-}
-
-impl<K> KeySize for [K]
-where
-    K: KeySize,
-{
-    // TODO(perf, bikeshedding): check that this optimized to `size_of::<K>() * self.len()`
-    // when `K` is a primitive.
-    fn key_size_in_bytes(&self) -> usize {
-        self.iter().map(|elt| elt.key_size_in_bytes()).sum()
-    }
-}
-
-impl KeySize for ArrayValue {
-    fn key_size_in_bytes(&self) -> usize {
-        match self {
-            ArrayValue::Sum(elts) => elts.key_size_in_bytes(),
-            ArrayValue::Product(elts) => elts.key_size_in_bytes(),
-            ArrayValue::Bool(elts) => elts.key_size_in_bytes(),
-            ArrayValue::I8(elts) => elts.key_size_in_bytes(),
-            ArrayValue::U8(elts) => elts.key_size_in_bytes(),
-            ArrayValue::I16(elts) => elts.key_size_in_bytes(),
-            ArrayValue::U16(elts) => elts.key_size_in_bytes(),
-            ArrayValue::I32(elts) => elts.key_size_in_bytes(),
-            ArrayValue::U32(elts) => elts.key_size_in_bytes(),
-            ArrayValue::I64(elts) => elts.key_size_in_bytes(),
-            ArrayValue::U64(elts) => elts.key_size_in_bytes(),
-            ArrayValue::I128(elts) => elts.key_size_in_bytes(),
-            ArrayValue::U128(elts) => elts.key_size_in_bytes(),
-            ArrayValue::I256(elts) => elts.key_size_in_bytes(),
-            ArrayValue::U256(elts) => elts.key_size_in_bytes(),
-            ArrayValue::F32(elts) => elts.key_size_in_bytes(),
-            ArrayValue::F64(elts) => elts.key_size_in_bytes(),
-            ArrayValue::String(elts) => elts.key_size_in_bytes(),
-            ArrayValue::Array(elts) => elts.key_size_in_bytes(),
-        }
-    }
-}
-
 /// A B-Tree based index on a set of [`ColId`]s of a table.
 #[derive(Debug, PartialEq, Eq)]
 pub struct BTreeIndex {
@@ -953,15 +836,7 @@ impl BTreeIndex {
     ///
     /// This method runs in constant time.
     ///
-    /// The key bytes of a value are defined depending on that value's type:
-    /// - Integer, float and boolean values take key bytes according to their [`std::mem::size_of`].
-    /// - Strings take key bytes equal to their length in bytes.
-    ///   No overhead is counted, unlike in the BFLATN or BSATN size.
-    /// - Sum values take 1 key byte for the tag, plus the key bytes of their active payload.
-    ///   Inactive variants and padding are not counted, unlike in the BFLATN size.
-    /// - Product values take key bytes equal to the sum of their elements' key bytes.
-    ///   Padding is not counted, unlike in the BFLATN size.
-    /// - Array values take key bytes equal to the sum of their elements' key bytes.
+    /// See the [`KeySize`] trait for more details on how this method computes its result.
     pub fn num_key_bytes(&self) -> u64 {
         self.num_key_bytes
     }
